fix(Grammar): allow arguments which are same as keywords
In this we start creating special KEYWORD_ tokens with higher priority
in order to parse these keywords using this token if they are followed
by a colon. We then use Transform stage to replace the KEYWORD_ type
with type NAME so as to provide following stages with coherance in
parsed token.